requirements.txt
pkg-resources==0.0.0
psycopg2==2.9.3
python-dotenv==0.20.0
Shapely==1.8.2
Unidecode==1.3.4
